feat: refactor and improve organiser dashboard
