+++ components/dashboard/organizer/password/index.tsx
@@ -1,40 +1,48 @@
-import { useState, useEffect } from 'react';
-import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { toast } from '@/hooks/use-toast';
-import { EventRole, EventVisibility } from '@/lib/types';
-import { supabase } from '@/lib/supabase';
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
+"use client";
+
+import { useState, useEffect } from "react";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { toast } from "@/hooks/use-toast";
+import { EventRole, EventVisibility } from "@/lib/types";
+import { supabase } from "@/lib/supabase";
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from '@/components/ui/select';
+} from "@/components/ui/select";
 
-interface RolePasswordSettingsProps {
+interface PasswordTabProps {
   eventId: string;
 }
 
-export function RolePasswordSettings({ eventId }: RolePasswordSettingsProps) {
-  const [judgePassword, setJudgePassword] = useState('');
-  const [organizerPassword, setOrganizerPassword] = useState('');
-  const [eventPassword, setEventPassword] = useState('');
+export function PasswordTab({ eventId }: PasswordTabProps) {
+  const [rolePasswords, setRolePasswords] = useState<Record<string, string>>(
+    {}
+  );
+  const [eventPassword, setEventPassword] = useState("");
   const [isLoading, setIsLoading] = useState(false);
   const [visibility, setVisibility] = useState<EventVisibility | null>(null);
+
+  // List of available roles from the EventRole enum, excluding Admin
+  const availableRoles = Object.values(EventRole).filter(
+    (role) => role !== EventRole.Admin
+  );
 
   useEffect(() => {
     // Fetch the current event visibility
     const fetchEventVisibility = async () => {
       const { data, error } = await supabase
-        .from('events')
-        .select('visibility')
-        .eq('event_id', eventId)
+        .from("events")
+        .select("visibility")
+        .eq("event_id", eventId)
         .single();
 
       if (error) {
-        console.error('Error fetching event visibility:', error);
+        console.error("Error fetching event visibility:", error);
       } else if (data) {
         setVisibility(data.visibility);
       }
@@ -43,15 +51,22 @@ export function RolePasswordSettings({ eventId }: RolePasswordSettingsProps) {
     fetchEventVisibility();
   }, [eventId]);
 
-  const handleSetPassword = async (role: EventRole) => {
+  const handlePasswordChange = (role: string, password: string) => {
+    setRolePasswords((prev) => ({
+      ...prev,
+      [role]: password,
+    }));
+  };
+
+  const handleSetPassword = async (role: string) => {
     try {
       setIsLoading(true);
-      const password = role === EventRole.Judge ? judgePassword : organizerPassword;
+      const password = rolePasswords[role] || "";
 
-      const response = await fetch('/api/roles/set-password', {
-        method: 'POST',
+      const response = await fetch("/api/roles/set-password", {
+        method: "POST",
         headers: {
-          'Content-Type': 'application/json',
+          "Content-Type": "application/json",
         },
         body: JSON.stringify({
           eventId,
@@ -66,7 +81,7 @@ export function RolePasswordSettings({ eventId }: RolePasswordSettingsProps) {
         toast({
           title: "Error",
           description: data.error || "Failed to set password",
-          variant: "destructive"
+          variant: "destructive",
         });
         return;
       }
@@ -77,17 +92,13 @@ export function RolePasswordSettings({ eventId }: RolePasswordSettingsProps) {
       });
 
       // Clear password input
-      if (role === EventRole.Judge) {
-        setJudgePassword('');
-      } else {
-        setOrganizerPassword('');
-      }
+      handlePasswordChange(role, "");
     } catch (error) {
-      console.error('Error setting password:', error);
+      console.error("Error setting password:", error);
       toast({
         title: "Error",
         description: "Failed to set password",
-        variant: "destructive"
+        variant: "destructive",
       });
     } finally {
       setIsLoading(false);
@@ -98,14 +109,14 @@ export function RolePasswordSettings({ eventId }: RolePasswordSettingsProps) {
     try {
       setIsLoading(true);
 
-      const response = await fetch('/api/roles/set-password', {
-        method: 'POST',
+      const response = await fetch("/api/roles/set-password", {
+        method: "POST",
         headers: {
-          'Content-Type': 'application/json',
+          "Content-Type": "application/json",
         },
         body: JSON.stringify({
           eventId,
-          role: 'event', // Special role type for event password
+          role: "event", // Special role type for event password
           password: eventPassword,
         }),
       });
@@ -116,7 +127,7 @@ export function RolePasswordSettings({ eventId }: RolePasswordSettingsProps) {
         toast({
           title: "Error",
           description: data.error || "Failed to set event password",
-          variant: "destructive"
+          variant: "destructive",
         });
         return;
       }
@@ -127,13 +138,13 @@ export function RolePasswordSettings({ eventId }: RolePasswordSettingsProps) {
       });
 
       // Clear password input
-      setEventPassword('');
+      setEventPassword("");
     } catch (error) {
-      console.error('Error setting event password:', error);
+      console.error("Error setting event password:", error);
       toast({
         title: "Error",
         description: "Failed to set event password",
-        variant: "destructive"
+        variant: "destructive",
       });
     } finally {
       setIsLoading(false);
@@ -143,37 +154,41 @@ export function RolePasswordSettings({ eventId }: RolePasswordSettingsProps) {
   const handleVisibilityChange = async (value: EventVisibility) => {
     try {
       setIsLoading(true);
-      
+
       const { error } = await supabase
-        .from('events')
+        .from("events")
         .update({ visibility: value })
-        .eq('event_id', eventId);
-      
+        .eq("event_id", eventId);
+
       if (error) {
         toast({
           title: "Error",
           description: "Failed to update event visibility",
-          variant: "destructive"
+          variant: "destructive",
         });
         return;
       }
-      
+
       setVisibility(value);
-      
+
       toast({
         title: "Success",
         description: "Event visibility updated successfully",
       });
     } catch (error) {
-      console.error('Error updating event visibility:', error);
+      console.error("Error updating event visibility:", error);
       toast({
         title: "Error",
         description: "Failed to update event visibility",
-        variant: "destructive"
+        variant: "destructive",
       });
     } finally {
       setIsLoading(false);
     }
+  };
+
+  const capitalizeFirstLetter = (string: string) => {
+    return string.charAt(0).toUpperCase() + string.slice(1);
   };
 
   return (
@@ -186,16 +201,20 @@ export function RolePasswordSettings({ eventId }: RolePasswordSettingsProps) {
           <div className="space-y-4">
             <div className="space-y-2">
               <label className="text-sm font-medium">Event Visibility</label>
-              <Select 
-                value={visibility || ""} 
-                onValueChange={(value) => handleVisibilityChange(value as EventVisibility)}
+              <Select
+                value={visibility || ""}
+                onValueChange={(value) =>
+                  handleVisibilityChange(value as EventVisibility)
+                }
               >
                 <SelectTrigger>
                   <SelectValue placeholder="Select visibility" />
                 </SelectTrigger>
                 <SelectContent>
                   <SelectItem value={EventVisibility.Public}>Public</SelectItem>
-                  <SelectItem value={EventVisibility.Private}>Private (Password Protected)</SelectItem>
+                  <SelectItem value={EventVisibility.Private}>
+                    Private (Password Protected)
+                  </SelectItem>
                   <SelectItem value={EventVisibility.Draft}>Draft</SelectItem>
                 </SelectContent>
               </Select>
@@ -203,7 +222,9 @@ export function RolePasswordSettings({ eventId }: RolePasswordSettingsProps) {
 
             {visibility === EventVisibility.Private && (
               <div className="space-y-2">
-                <label className="text-sm font-medium">Event Password (for Private events)</label>
+                <label className="text-sm font-medium">
+                  Event Password (for Private events)
+                </label>
                 <div className="flex gap-2">
                   <Input
                     type="password"
@@ -233,44 +254,33 @@ export function RolePasswordSettings({ eventId }: RolePasswordSettingsProps) {
         </CardHeader>
         <CardContent>
           <div className="space-y-4">
-            <div className="space-y-2">
-              <label className="text-sm font-medium">Judge Password</label>
-              <div className="flex gap-2">
-                <Input
-                  type="password"
-                  value={judgePassword}
-                  onChange={(e) => setJudgePassword(e.target.value)}
-                  placeholder="Set judge password"
-                />
-                <Button
-                  onClick={() => handleSetPassword(EventRole.Judge)}
-                  disabled={isLoading || !judgePassword}
-                >
-                  Save
-                </Button>
+            {availableRoles.map((role) => (
+              <div key={role} className="space-y-2">
+                <label className="text-sm font-medium">
+                  {capitalizeFirstLetter(role)} Password
+                </label>
+                <div className="flex gap-2">
+                  <Input
+                    type="password"
+                    value={rolePasswords[role] || ""}
+                    onChange={(e) => handlePasswordChange(role, e.target.value)}
+                    placeholder={`Set ${role} password`}
+                  />
+                  <Button
+                    onClick={() => handleSetPassword(role)}
+                    disabled={isLoading || !rolePasswords[role]}
+                  >
+                    Save
+                  </Button>
+                </div>
+                <p className="text-xs text-gray-500">
+                  This password will be used by {role}s to join the event.
+                </p>
               </div>
-            </div>
-
-            <div className="space-y-2">
-              <label className="text-sm font-medium">Organizer Password</label>
-              <div className="flex gap-2">
-                <Input
-                  type="password"
-                  value={organizerPassword}
-                  onChange={(e) => setOrganizerPassword(e.target.value)}
-                  placeholder="Set organizer password"
-                />
-                <Button
-                  onClick={() => handleSetPassword(EventRole.Organizer)}
-                  disabled={isLoading || !organizerPassword}
-                >
-                  Save
-                </Button>
-              </div>
-            </div>
+            ))}
           </div>
         </CardContent>
       </Card>
     </div>
   );
-} 
+}
